requirements.txt
PyGithub==2.1.1
Pillow==10.1.0
matplotlib==3.8.2
requests==2.31.0
